refactor: replace Session-DI with ServerRepository Port in servers application layer (Resolves #285)
## Summary
- Remove the dead `Callable[[Session], ServerRepository]` factory from
  `MinecraftServerManager` (no longer consumed since #272), and drop
  the late-bound wiring in `app/main.py`.
- Move the SQLAlchemy-direct legacy helpers `ServerJarService` and
  `ServerDatabaseService` out of `app/servers/application/service.py`
  into `app/servers/adapters/_legacy_helpers.py`. The application
  module re-exports them for the existing unit-test imports.
- Lift the `Session` annotation in `service.py` under `TYPE_CHECKING`
  so `app/servers/application/` carries zero runtime `sqlalchemy`
  imports (ARCHITECTURE §4.2).
- Convert the legacy `ServerService._is_version_supported_db` body to
  a thin passthrough into a new `is_version_supported_db_legacy(...)`
  adapter helper (so the service module stops constructing
  `SqlAlchemyVersionRepository` directly).
- Replace the in-method imports of `make_server_repository_from_session`
  (which pulled `api/` into `application/`) with a small
  `_require_server_repo(db)` helper that returns the constructor-
  injected `ServerRepository` Port and raises a clear error when the
  service was built without DI.

## ARCHITECTURE §4.2
After this change, `app/servers/application/*.py` has zero runtime
`from sqlalchemy ...` imports and zero runtime imports from
`app/servers/api/`. The only remaining `Session` reference is a
TYPE_CHECKING-only annotation on legacy passthrough methods.

## Test plan
- `uv run pytest tests/unit -m "not slow"` — 1107 passed, 5 skipped.
- `uv run pytest tests/integration -m "not slow"` — 427 passed, 5
  skipped.
- `just test` — 1719 passed, 10 skipped (full suite incl. slow).
- `uv run ruff check app/` clean.
- `uv run pre-commit run --files <touched>` clean.
